(Flatmap) Python binding for the computation of streamlines intersections [NSETM-1035]
This change implements the function compute_streamlines_intersections,
to be used by atlas-building-tools for flattening 3D brain regions with a laminar
structure, i.e., layered regions such as the rat or the mouse isocortex.

This function is implemented as a C++-python binding for efficiency reasons:
trials with python have yielded unreasonnable runtimes,
e.g., days or weeks, in the case of the Paxinos-Watson rat S1 region.

The inputs are:
 * `layers`, a uint8 numpy array of shape (W, H, D) whose voxels are labeled
by layer indices; W, H and D are the integer dimensions of the array.
 * `vector_field`, a float32 numpy array of shape (W, H, D, 3) to be interpreted
as a 3D unit vector field over the `layers` domain.
 * the offset and the voxel dimensions of the voxcell.VoxelData object
initially associated to `layers`; they define the transformation
from index space to world space.
 * the layer indices of two layers sharing a common boundary

The surface boundary of the two specified layers is intersected with
streamlines which follow the `vector_field` stream. For each non-void voxel
of `layers`, the algorithm draws the streamline passing through it and
computes (an approximation of) its intersection point with the surface
boundary. This yields a map which assigning 3D points to voxels.

The output is a float32 numpy array of shape (W, H, D, 3), to be read
as the aforementioned map assigning 3D points to the voxels of `layers`.

Note: the companion commit in atlas-building-tools is
https://bbpcode.epfl.ch/code/#/c/50053/

Change-Id: Ica4d7af485fe77f3fcff27190b85af60fd1328a9